Remove bobthecow/Ruler
Unfortunately this library does not appear to be actively maintained anymore. The contributor _has_ responded to one comment or another here and there, but it's been very infrequent. There are issues and PRs stacking up, and contributors in the comments wondering if the library has been abandoned.

Additionally, 1) the library is supposedly built for PHP5 and not PHP7, and 2) the library is not findable on Packagist, making it unnecessarily difficult to install it using Composer without a fork.

Therefore I sadly but respectfully submit this library for removal from the Awesome list.